Added option to only create a view of hinted tracks rather than a new slim track collection.
A new option was added to only create a view of the  tracks which are to
be kept. In this mode it is assumed that the slimming tool will only
set persistification hints but not create new slim tracks.

Also fixed endreq compiler warning by using ATH_MSG macros or replacing
endreq with endmsg. (TrkTrackSlimmer-01-00-01-01)

        * Ported to DataHandles
	* I have preserved ability to process multiple collections,
	* but in a multithreaded environment it would be better to schedule
	* multiple instances of algorithms with one collection each
	* tagged TrkTrackSlimmer-01-00-01


Former-commit-id: af76f5737f224f2eed4c488deea67a1fb498721c