use python wrapper for gitlab API to fetch changes of merge request
Similar to the rest of the CI scripts, we can make use of the
python-gitlab module which is a neat wrapper around the Gitlab API.
Previously there was a bug when fetching the changes introduced by
a merge request. This is fixed in python-gitlab v0.17. Therefore,
the manual workaround using HTTP requests directly is not needed
any longer.


Former-commit-id: 31762bcf4f36217ff1d24e065f2f7ce682c28123