limit the number of affected packages mentioned in MR discussion
For repository-wide changes, many packages are affected. If the list
of affected packages becomes very long, the automatic comment from
the atlasbot does not really help, but only makes following the
discussion on the Gitlab web interface more difficult. Therefore,
the list of affected packages is only printed for at most 20 packages.

Closes ATLINFR-1488


Former-commit-id: c5080e62274c294ccd3405c10f8b31e3b4072639