script for adding comments to MR needed to publish links to logfiles
Due to the problem of creating duplicate CI jobs when a MR is updated
while the initial CI job is still running, a mechanism has been put
in place to cancel the second (duplicate) job. This leads to a failed
Jenkins CI job. Publishing the build status using the functionality
of the jenkins-gitlab plugin would cause major confusion since the
developer would quickly get a 'failed build' comment from the aborted
job and later the actual build result from the initial CI job (which
can take up to 5h to complete). Therefore, this functionality is not
used but a comment with the build status (and links to the logfiles)
is added manually to the Gitlab MR discussion.


Former-commit-id: babe005cce971347d5a0298e544545fc018d1a4b