domain labels for already accepted MRs are still useful
In the case of short-circuited MRs (skipping the review process),
a MR can be closed before the CI job for assigning the Gitlab labels
is run. In this case, domain labels should still be assigned to
facilitate future searches for MRs by domain. Review labels are still
added in case a post-facto review is desired.


Former-commit-id: 7b12fb512e0d7e5811a2c83e01c457f7c74277d4