Improved build helper scripts' robustness and functionality
In the externals build script if an install of AthenaExternals and
Gaudi is found then the script will exit, unless -f is specified. This
allows the script to be called multiple times safely, e.g., by the
continuous integration system. Specifically the check is done against
installs of AthenaExternals and Gaudi, so failed builds should not
cause the script to think that there is a valid installation.

In the main build script, factorise out the environment setup piece to
build_env.sh, allowing it to be easily sourced when someone wants to
manage the build themselvs or invoke the build commands in special
ways.

Support options in the main build script to control which pieces of
the build are done (cmake, make, install, cpack) allowing more
detailed control (e.g., the continuous integration system certainly
doesn't care about making RPMs).


Former-commit-id: 6412405cafcf41f8234da6db185107b4c3543a78