Fixed MadGraph_i for linking errors. It was missing the UPEVNT() subroutine.
Now, as far as I understand, linking in pythia6_dummy may have been a problem
in the end. As it could very well be that the intention was very much to only
find this subroutine at runtime, in some different library. In that case the
package will have to be modified to give up on using -Wl,--no-undefined for
its library.


Former-commit-id: 30aa1b483851bcc85907d6dc44d2f69ff915d6ac