Fixed the build problems of Tauola_i.
Removed the getaddr.cxx file from the package, as it's provided by
GeneratorFortranCommon by now.

Updated the C++ code to compile without warnings against new versions
of Gaudi.

Enabled the -DATLAS_CMAKE build definition for the main library, as the fortran code was
set up a long time ago to behave a bit differently with this enabled.
And don't call on non-existent subroutines.


Former-commit-id: 7fa3e4d22b1e0de63e7da2215cdb303f98136125