indicate result of external build in CI status summary
A failure in building the externals marks a CI job as failed and
cancels all subsequent stages. This would result in MR comment by
atlasbot saying that the CI job failed but none of the stages were
actually run. This can be quite confusing. Adding the external build
stage makes it more clear at which step something went wrong. Even
if it is not clear for a normal developer why building the externals
is affecting their MR, it gives valuable information to the experts.


Former-commit-id: f4f2106b88144389745932f16c1c2b7c93bf33a6