Fix coverity defect 13834 (deref after null check)
Previously this method would print a VERBOSE message after being unable to
dynamic_cast a surface object to testPSf, then proceed to dereference testPSf.
Coverity is not happy about that.  Now a similar VERBOSE message will be
printed, but the pointer will not be deref'd.

Also change a couple of pointer initializers, 0 -> nullptr


Former-commit-id: c5806289425bdb5b257ea89f298272746c2cdb49