Pulled in AthenaKernel-00-60-06-01 and CLIDComps-00-06-24 from SVN.
These are the versions of the packages used in AthAnalysisBase-2.6.X at the
moment. As they fix exactly the same race condition that was hitting this
branch in the previous build attempt. Where the build would try to execute
the genCLIDDB executable before all of its dependencies would be ready.

The two tags are quite a bit newer than the ones used in 21.0 at the moment,
but based on the experience with AthAnalysisBase, they should hopefully
work without any issues in this branch.


Former-commit-id: 14b1b5a2df3c1dd7726c9c7576118a9cfff880ae