Pulled in JetTagTools-01-01-27 from SVN.
This was needed to adapt to the updated xAOD::BTagging class.

I also had to add an explicit link of JetTagTools to VxSecVertex. Something
is pulling in a public dependency on that library into the component library,
but I was unable to figure out what. :-( Some piece of code seems to call
Trk::VxSecVertexInfo::clone() somewhere that's visible to JetTagTools_entries.cxx.
But however I looked, I couldn't find out what piece of code that is. So
to be able to move on, for now I just declared this extra dependency. Even
though the component library definitely doesn't depend on VxSecVertex
directly.

Former-commit-id: a4aa01363e5d7be6bd33044b94035998803fdff3