Adding dynamic aux store variables to SharedWriter
Up to now streaming did not include any dynamic auxiliary store variables:
This change introduces two new private member functions to the AthenaPoolCnvSvc
to access the aux store and send it to the streaming tool and to receive the
aux store from the streaming tool. These methods are used for reading and writing.


Former-commit-id: 533f66aa9024b2b17c9931721fb73cc45a842e25