Added the external setup files that were previously in the AthenaExternals project into the Athena project's configuration. This is to allow us to fine-tune which externals we want to use in each branch.
Updated the project configuration of Athena to make use of the new files.

Updated the Athena project to build against the latest hash of the
atlasexternals repository. While the update is being tested. This will
have to be changed to an annoted tag once the build is working
correctly.

Updated the package filters file of the Athena project to work correctly
with the latest version of the AtlasCMake code.


Former-commit-id: 2038615812fb15d3115d642f17b8970dbc29496e