Had to update the 21.0 commit to use the newer versions of Rivet and YODA that were by now configured for the master branch.
Made sure that the generator externals set up by the Athena project would
be in sync with what some of the generator packages requested explicitly.

Then I removed the explicit requests from the generator packages, to only
have the version declarations in a single place.

Also had to add a configuration file for Herwig3.


Former-commit-id: 3022153185eb418201ed7de811a1757a0b14afae