Use again the pixel cluster NN during ROT creation.
The GlobalChi2Fitter does not seem to handle brem fits properly when
fitting PrepRawData rather than ROTs. Therefore, brem fits are again
performed via the code path which starts from ROTs which requires that
the ROTs already contain the enhancements from the Pixel cluster NN.
This is a temproary fix for ATLASRECTS-3949. (InDetRecExample-03-01-05-06)


Former-commit-id: ba5bf93ae466d3b08f704d7bb87537b6975f92e9