Added a dummy CMakeLists.txt file into the Build directory.
This is needed to make the CI system assign a reasonable label to merge requests modifying
files in this directory.

At the same time updated the Athena package_filters.txt file to ignore both the CI and
the Build directories. Not having the project ignore them was not leading to any compilation
problems, but it made both directories show up in the generated packages.txt file. Which
really just tries to confuse NICOS.


Former-commit-id: d2ef8801f975f97b8492262dfba6bb300bad56e6