Fixed the package configuration of PixelConditionsServices.
First of all, introduced an interface library for the headers that the package
provides to the outside world. So that client packages could correctly pick
up those headers by linking against this interface library.

Also updated the component library's configuration to link against the newly
created PixelGeoModelLib library. To resolve the missing symbols that showed
up with GCC 6.2 in debug mode.


Former-commit-id: 10332ae95d77357008356593c8c1611e9218015f