Re-organised the compilation rules of PixelCalibAlgs.
The build of this package was set up quite inefficiently. And it was using
very non-standard library names as well. The package now instead builds
libraries with standard dual-use library names, and links them to each
other. (So that source files would only need to be compiled once.)

The previous non-standard library name (PixelOfflineCalib) was only used
in the PixelCalibAlgs package and in PixelRTT. Updated both of them to
use the new (PixelCalibAlgsLib) installed library name instead.

While doing this, simplified the configuration of the PixelRTT package
a bit.


Former-commit-id: 188f4ba28e82f5865466352591b3d6dfe31e41a6