Fixes for Qball and Monopole simulation configuration.
Monopole and Qball simulation jobs had not been tested for a while.
Production tests in 20.3.7.3 revealed a number of configuration issues.
These are fixed by the changes in this commit. See ATLASSIM-3251.
There are still some run-time issues for Monopole simulation jobs,
which are being followed up on the above ticket.


Former-commit-id: 077511dc0f7a8807f5bef84d259b1c6c4824be47