Fixed the linking problems of TrkParticleCreator by linking its main library explicitly against PixelGeoModelLib.
Unfortunately some trigger packages do include the header file of the
package, even though it should not be usable for anything. (Its symbols
are not available outside of the package.) Still, to keep those packages
working, I introduced an interface library for the one header for now.

That should just be removed from the configuration once the clients of
the package are removed.


Former-commit-id: 7bc6132787ab0af06f4b2b97f8a499a131fe45b3