port CI changes into 21.0
The python scripts used by the CI system have evolved a bit. While
the current version of the CI scripts in 21.0 are still functional
they exhibit a different subtly different behavior (most importantly
the assignment of review labels). To simplify the life of the MR
review shifters and to streamline the CI setup, this MR brings in
the changes to the CI scripts from the master branch.


Former-commit-id: 29ff962eed9d7851fce8e404329c19503d89a63a