provide script for sweeping merge requests between branches
This scripts provides a convenient way of cherry-picking merge
commits from a source branch to a list of target branches, The
source branch must have a file 'CI/config.yaml' which defines
a list of target branches with the name 'sweep-targets'. This
script expects in addition a definition of a time interval.
Only merge commits during this time interval will be cherry-
picked.


Former-commit-id: 1c86014ba568c7e82a2c84635ebdd07f31968311