Enable Concurrent/Sequential Sequences in MT
Add Property "Sequential" to AthSequences (default=false), which
indicates to the Scheduler how to process the Algorithms in it. If
Sequential=False, processing is unchanged from current behavior,
and only data flow rules will determine the scheduling. If Sequential=True,
the Scheduler will enforce strict ordering of the Algorithms, meaning that
one must terminate before the next can start. The Algorithms can still execute
in different threads, and other Algorithms may be scheduled between them.

This Property has no impact in Serial running, and requires the Properties
useGraphFlowManagement and DataFlowManagerNext of the ForwardSchduler to be
set "True".


Former-commit-id: 7782ffe6c734d36cfe071092e31a11709146e8ca