propagate number of shell execution errors
Even if we do not to fail immediately on errors when building Gaudi
or ATLAS externals, it is still important for the CI system to know
whether some errors occured. By not failing on error but reporting
the total number of errors, the calling scripts can decide how to
deal with this in specific situations (e.g. CI build vs nightly).


Former-commit-id: 136d2f165ed40ec1822042f09d3449f1857480bc