AthContainers+CxxUtils: Fix ubsan warnings.
A (deliberate) reinterpret_cast in copyAuxStoreThinned triggers a warning
from the undefined behavior sanitizer in gcc6.  For now, disable warnings
for this particular function.  This is done with a gcc-specific function
attribute, so add a macro in CxxUtils to make this visible only for gcc6
or later.

Also fix another ubsan warning in one of the tests.

Former-commit-id: 2627592a7d8ce70a609c4c9f01af648220eee4fa