SGTools+AthContainers: undefined behavior sanitizer fixes
With gcc6, we see warnings from the undefined behavior sanitizer from BaseInfo
when we run things like genCLIDDB.

In a couple places, we call baseinfo() to get the BaseInfo structure
initialized.  This returns a const reference to something that is created
in a static constructor.  However, if the baseinfo() call itself is
triggered from a static constructor, then we may end up returning
a null for the reference, if the constructors run in the wrong order.

This is harmless in itself, as the returned reference is ignored; however,
it (properly) causes a storm of ubsan warnings.  Fix this by changing
the baseinfo() calls to a new maybeInit() method which does the same
thing but does not return anything.

Former-commit-id: f0ac0f28aebbe6fd8f413558b176fca5714e951b