Fixed the configuration of TrigAnalysisExamples.
Removed packages from the atlas_depends_on_subdirs(...) call that make no
difference/sense for the CMake build.

Moved the source of TrigAnalysisExApp into a separate directory, such that
it would not be built into the component library of the package.

Updated the build configuration to link TrigAnalysisExApp correctly
against TrigDecisionTool.


Former-commit-id: 3e15029b0830288b1ea8ad3299a1ea5943bec8a4