First commit in the process of splitting egamma into more algs
A new directory(package), egammaAlgs, was created, into which the
main egamma algorithms are to be placed. It started out as a copy
of egammaRec, with EMBremCollectionBuilder and EMVertexBuilder
added from egammaTools. egammaRec remains as a python-only package.
The current code compiles, but the python configuration updates
have not yet been made, so it doesn't run.


Former-commit-id: c2be47ad7294aac992fe2e6c29d61e36057a8870