Adding function for determining if a particle has a B-hadron or b-quark in its history and moving some functions to use the MCUtils helper functions
The first function is a request of groups trying to understand cascade
decays, so that they can call a function on their lepton or c-hadron.
The second is a long-requested move of some helper functions to the
MCUtils central helpers.  This doesn't move all the functions to use
these helpers, but it is a start.


Former-commit-id: fed9577e6d55de0bd6d5a6c3c5fd3801c56a9116