Simplify varhandle initialization for handles that are possibly ignored.
If a varhandle is used, we must call initialize on it; if it is not used,
the key must be cleared.  Combine these by extending initialize() to take
a flag.  If true, we proceed as usual; if false, we clear the key and
return success without doing anything else.

Also some small cleanups to tests.



Former-commit-id: 26c5231376d3f28b38929321ad204372fcd37916