Fix linking of TrigT1Monitoring in dbg build.
TrigT1Monitoring directly references L1CaloCondSvc from TrigT1CaloCondSvc,
which is a class with out-of-line virtual functions.  Thus, it should
link against the library where these functions are defined in order
to see the typeinfo (which it needs) and vtable (which it doesn't
really need).  However, TrigT1CaloCondSvc was built only as a component
library.

In a previous MR, TrigT1CaloCondSvc was changed to a dual use library.
We now link to this library to resolve this problem.    

A better solution would be to rework L1CaloCondSvc in terms of an abstract
interface, so that linking against TrigT1CaloCondSvc would not be
necessary.  However, L1CaloCondSvc has non-trivial templated functions
accessing the class internals, so this is not trivial.
Leaving that for someone else to enjoy doing...


Former-commit-id: ec918627ee2b0e3769e8aaab4bf1b2391a54bc7d