Fix linking of TrigT1CaloCalibTools in dbg build.
TrigT1CaloCalibTools directly references L1CaloCondSvc from TrigT1CaloCondSvc,
which is a class with out-of-line virtual functions.  Thus, it should
link against the library where these functions are defined in order
to see the typeinfo (which it needs) and vtable (which it doesn't
really need).  However, TrigT1CaloCondSvc was built only as a component
library.

In a previous MR, TrigT1CaloCondSvc was changed to dual use library.
This changes TrigT1CaloCalibTools to link against it, resolving the error.

A better solution would be to rework L1CaloCondSvc in terms of an abstract
interface, so that linking against TrigT1CaloCondSvc would not be
necessary.  However, L1CaloCondSvc has non-trivial templated functions
accessing the class internals, so this is not trivial.
Leaving that for someone else to enjoy doing...


Former-commit-id: ac40d14f1e5a82acfcc357eb1b18d27ae9513998