CaloTPCnv: Adjust use of DataPool to work properly with locking.
Upcoming changes to DataPool will have the DataPool object take out a lock.
This is needed for proper MT operation, but means that creating two
instances with the same type in the same thread will deadlock.
This wouldn't usually happen, but it did occur in one of the unit tests.



Former-commit-id: f420b7a1753ed16a24246db84b56c26c21c1dd7e