StoreGate: Introduce a separate lock for the remaps table, to prevent deadlocks.
A deadlock was observed involving the SG and pool converter locks.
One thread was holding the SG lock via retrieve and trying to acquire
the pool converter lock.  Another thread was holding the pool converter
lock and trying to acquire the SG lock to access the remap table.

Resolve this for now by introducing a separate lock for the SG remap table.


Former-commit-id: dae94c9f652086ca7733d49996ce58422cc6f834