Addressing comments from original commit, more code cleanup
Moving more of the code to use helpers from MCUtils.  In the future we
will need to ask for a couple of helpers to be added.  We should also
continue to migrate to using the xAOD::TruthParticle function helpers
instead of the static functions, once they are all available.  For now,
I've done some more migration, though there is a lot of code needing
work.


Former-commit-id: 335f6b8aee100b775250551e2dcd002714170adf