ParticleEventTPCnv: Fix tests to avoid DataPool deadlocks.
Upcoming MT fixes for AthAllocators will have DataPool holding a lock
on the underlying allocator.  This leads to a deadlock if there are
two live instances of DataPool for the same type in the same thread.
This shouldn't normally happen, but it was in a few unit tests.
Cleaning these up.



Former-commit-id: ce353ca4fc7ba7b60916c1375da5829cd739a97a