GeneratorObjectsTPCnv: Fix tests to avoid DataPool deadlocks.
Upcoming MT fixes for AthAllocators will have DataPool holding a lock
on the underlying allocator.  This leads to a deadlock if there are
two live instances of DataPool for the same type in the same thread.
This implies that a DataPool instance cannot be copied.
Fix a couple places where GeneratorObjectsTPCnv was doing this.



Former-commit-id: 48a38c08ab190a24452a1a799849bc996019a948