Made the AtlasPyFwdBwdPorts package build successfully.
For this I had to add a few extra source files that got removed in the
SVN->Git migration, and just disabled a few packages that I found too large
to include into Git. (And which don't seem to be desperately needed.)

Had to resolve some conflicts in the AtlasPyFwdBwdPorts/CMakeLists.txt file.


Former-commit-id: cf5d74eba6d03dfb5f04b269ce20e7f76c1505e4