fix linking for standalone L1Topo executables
Details:
These executables are used to test the l1topo simulation in standalone.
However, the previous CMakeLists.txt would not link them correctly
against L1TopoAlgorithms, causing the runtime error
```
Caught exception: AlgFactory: trying to instantiate algorithm of type ClusterNoSort with name EMall, but it has not been registered.
```

Note that this problem was not observed in the athena tests because
with athenaHLT.py the libraries are loaded in some smarter way.
Thank you @akraszna for providing the solution and explanation:

"The problem here (I think) is that GCC doesn't recognise that the
executable actually needs to be linked against all the libraries that
were specified. Since we use the link flag -Wl,--as-needed when
linking everything in ATLAS, if the compiler decides that the
TrigConfTopoStandAlone executable doesn't really need L1TopoCoreSim,
it just silently doesn't actually link the executable against it.

With the atlas_disable_as_needed() call you disable the usage of this
linker flag for the entire package. Which I think is fine for this
package on the whole."


Former-commit-id: 58a3abc65b995de7b7880d1ee410f4491dda013d