CxxUtils: Add ubsan_suppress function.
gcc6.2 expands the warnings from the undefined behavior sanitizer.
However, there are some false positives, and also real warnings coming
from external code that we don't want to change.

Some ubsan warnings can be suppressed through a file read at runtime,
not all can.  (It looks like gcc7 will be better in this regard, but i haven't
actually tested it.)

However, ubsan will not report a warning for a given source location
more than once.  So the hacky way of suppressing ubsan warnings is
to actually trigger the warning with standard error (fd 2) temporarily
redirected to /dev/null.

This change provides a function to do that.

I hope that this can be removed once we switch to gcc7.



Former-commit-id: 3044bda42161798f12dde4ebebffcc5f917c9b54