Fix potential deadlock between AthenaPoolCnvSvc and StoreGate.
In MT, we can have one thead calling StoreGate retrieve to read data.
This will take out first the SG lock for the event slot and then
the AthenaPoolCnvSvc lock.

Simultaneously, another thread running in the same event slot can dereference
a DataProxy causing a read.  This will take out the AthenaPoolCnvSvc lock.
However, it is possible for the I/O process to call back into StoreGate,
for example to set up a DataProxy for an ElementLink being read.
This will try to acquire the SG lock for the event slot, causing a deadlock.

Resolve this by having DataProxy::readData take out the SG lock before
calling the converter.  This is done by introducing a createObj method
in IProxyDict.


Former-commit-id: 60cd844d51515801c67520541790cfebe41aaf41