MuonGeoModel: Fix undefined behavior sanitizer warnings.
We were getting warnings about casting to a SupComponent, where the
object was not actually a SupComponent.  But in fact, objects of
SupComponent are never created.  The only difference between SupComponent
and its base class Standard Component is the addition of a few methods
that use public class data.  So we can just as well not have SupComponent
derive from StandardComponent and change those methods to statics,
taking a StandardComponent as input.

Also removed Sup.h/Sup.cxx, which is not actually used anywhere.



Former-commit-id: 264b4f3c563fc41697829b9252aa4338ff28fc49