MuonGeoModel: Fix undefined behavior sanitizer warnings.
We were getting warnings about casting to a SpacerComponent, where the
object was not actually a SpaComponent.  But in fact, objects of
SpacerComponent are never created.  SpaComponent and SpaComponent
are identical except for the name, so resolve this error by getting
rid of SpacerComponent (which never actually had any instances created)
and using SpaComponent instead.



Former-commit-id: fbfb4047a3cadf7b3aa2c764c3e855fae739e136