Now installing the ubsan.supp file already during the CMake configuration step. Since it may be needed during the build already before it would be installed with the standard atlas_install_runtime(...) function.
This creates a bit of a dependence in this package on the internals of
AtlasCMake, but this still seemed like the best approach for fixing the
nightly build.

I also moved the configured CxxUtilsEnvironment.cmake file into the
CMakeFiles/ directory in the binary directory. As I'll try to move more
and more code there. (To make possible in-source builds a bit easier
to manage.)


Former-commit-id: 715b7b9ce11977650e12d116a822a576c4d46705