Switch TopoUseBitwise to default 'true' (ATR-14949)
Details:

This change will swith the l1topo simulation run at HLT to use the
bitwise implementation of the algorithms (for those algorithms that do
have one).
This decision was discussed in ATR-14949 and the functions
in question have been validated in the reprocessing ATR-15680.

No change in counters is expected. The tests run fine.


Former-commit-id: bff74042a92a621f201044555e9ccc949eac234f