ATR-15815 disable the NtElem tool - a HLT cost monitoring sub-tool.
It collects a summary of all Trigger Elements in the event and exports this to the cost monitoring ntuple. This can in theory be used to re-form the navigation graph offline, but in reality there has never been any real use to do this in run 2.

Disabling this tool will save CPU in collecting the data and RAW disk space.


Former-commit-id: 00b8fe2b8b35dc3fbb0625f9d7b117a17180645c