Add new High Pt Minimum Bias PileUpType
Having a separate PileUpType for high pt minbias allows
us to distinguish it from the low pt minbias. This will
be used for the new McEventCollection structure. ATLASSIM-2430

Using the new PileUpTimeEventIndex constructor, which takes the
PileUpType, in the PileUpMergeSvc will allow us to query the
PileUpType of SimHits within the PileUpTools. ATLASSIM-2430