ATR-14570 Rework how Trigger Conf data are exported P1->T0 as part of HLT cost monitoring
This package is used at P1 and in T0 - hence this commit needs to go in 21.0 and 21.1. At this moment both projects are being built from 21.0 branch?

In P1 I changed it so we do not export trigger configuration data to the calibration_CostMonitoring byte stream.

At T0 I changed the unpacking such that if no calibration data are present in the byte stream - the data are instead read from COOL.

The subsequent export to ntuple should then proceed as before.

I have tried to test this as best I can offline but this is not easy. I have tested the disabling of the config writing and the loading
of configs from the DB indipendently. I was unable to test post-processing as a python based setup to prevent writing the config data has
no DB keys to fetch "at T0" and a database setup using M13 keys doesn't have the do-not-write-config code.

Also: comply with naming convention, fix depricated outputLevel()

Former-commit-id: f6681051f13239ee6e74cfe79faf71eff35de4e0