Add override environment variables for AtlasExternals and Gaudi checkouts
Add some environment variables with a standard pattern of
External_URL and External_REF that control the repository
URL and refrence used in the checkout scripts. This makes
it much easier to test changes to these repositories
before they are merged into the main repos. The special
value "current" for the URL signals no clone or checkout
and leaves the current checkout in place, so that even
uncommitted changes can be tested.


Former-commit-id: 061edbfd49e1c4a58724ef6817f6cf0e3ed84652