Type std::vector< std::vector< ElementLink< IParticleContainer > > > now suppported variable type for trigger jets in xAODJet EDM.
The ghost association of FTK tracks to jets in the jet trigger produced a dynamix complex variable of this type, not supported by the ByteStreamAuxContainer. Support was therefore implemented
in xAODJet. A new version of the JetTrigAuxContainer was created with additional attribute "GhostTrack", along with the necessary schema evolution changes.
Changes pulled in from svn tags xAODJet-00-01-40-03 and xAODJetAthenaPool-00-00-09-01.

Offline tools now support the ghost association of FTK tracks in trigger jet reconstruction.


Former-commit-id: 7321e0dbb8a0741bea2be90e7af391724443efca