Correctly assign ME tracks for MuGirl muons
If the standard combined fit fails to find a muon, it will probably be found by
the inside-out MuGirl algorithm.  However, the MS track originally found by the MS
track-finding algorithm will not be associated to this muon, as MuGirl starts with
segments in the MS rather than tracks.  Since this track represents our best
attempt at finding the track in the MS, and may well have a different MS hit content
than the MuGirl track, experts might still want access to it for debugging purposes,
so it is now added to the muon as the MSOnlyExtrapolatedTrack.  Note that this is
only for expert usage, analyzers will never use this track.

Furthermore, this also resolves an issue with ME track truth-matching: by adding this
track to the MS-only extrapolated tracks container, we avoid the possibility that
this track ends up being the truth-matched ME track rather than the ME track that
actually belongs to the MuGirl muon (which can happen if the MS track has more MS hits
than the ME track).


Former-commit-id: 87a1cf9dc408273ffa8b4d5c6eb8a3e700c000f5