LArMonTools: Avoid calling methods through a null pointer.
Methods of LArCoverageFCALBins were being called through a null pointer.
This worked previously because none of the methods accessed any data members.
However, this is undefined, and is now causing a crash when the undefined
behavior sanitizer is used.

Change LArCoverageFCalBins so that all methods are static, and get rid of
the pointer to it held by LArCoverage (which was never set to anything
non-null).



Former-commit-id: 7fd0495fe3e63cd23d78e80e0ca5b30a6001eb2f