D3PDMakerConfig: Fix sequencing of maker algorithms for MT.
The egamma and physics d3pd makers were scheduled as streams,
not algorithms.  This is not compatible with the current Gaudi, which
distributes incidents from an algorithm.  In that case, the maker
algorithm runs after the EndEvent incident, which causes some tools
to malfunction.

Changing the makers to run as algorithms instead of streams
fixes it.



Former-commit-id: 8bcb5c0506a3dfdd13de10a24d285a387305b662