bugfix: move decision.setBit outside for(tob1) loop
Details:
This algorithm should require that NONE of the tob1's are matched to
any of the tob2's. We have observed events in
`NOT-0MATCH-TAU30si2-EMall` in which there is one TAU30i matched to an
EM and another one which is not matched. The second one has
`accept=true` and sets the bit to 'pass, although it shouldn't (the
firmware does not accept those events).

Former-commit-id: 1374aa0cb20af8d4dbf14f85bd0299fd05ffe7a0