Bug fix for CombinerOnly and changed to PullCut to match offline configuration.
PullCut, and PullCutPhi changed for TMEF_OutwardsMuonTrackCleaner and TMEF_TrackCleaner to match the offline configuration.
Fixed a bug in the caching of CombinerOnly section of TrigMuSuperEF which was duplicating muons in the trigger instead of detecting two individual muons with different Pt, eta and phi.
JIRA ticket number: ATR-11240
Test results in: /afs/cern.ch/user/d/dlack/public/muonTriggerTests


Former-commit-id: f6619db0d7754dac1347feb92b4e5754ba436c19