MSVertexUtils: Const fixes
A long-standing bug in DataVector meaens that it is possible to get
a non-const pointer from a const_iterator.  This change is working
to make it possible to finally fix this, by fixing another package
that was taking advantage of this bug.

Also make accessor methods const, and don't pass large objects by value.




Former-commit-id: 599b0ba83bfe8f144b98eeda54dbf39cabe9e7e3