Add functionality to set jet build tool ptmin and ptminfilter variables.
These variables make cuts on jets reconstructed by the tool. The ability to make such cuts
is limited to j0 jet chains. The new values in SignaturesDict are rcu0, rcu5 (to set the ptmin
cuts ot 0 and 5.0 GeV restpectively) and rcc0, rcc5 to set the the ptminfilter cut to the same values.
IT is also possible to request the default pt cut, which is currently set at 7.0 GeV

ChainConfigMaker extracts the rcc and rcu values ot from the dictionary, and sets them in the jet_params object.
They are picked up from here by AlgFactory. AlgFactory also does the conversion of the cuts to MeV


Former-commit-id: 20e46bfb4a4f1ea3d5065d850b717969ab6c939b