adding back files skipped during git migration (ATLASRECTS-4009)
During the git migration non code files bigger than 100kB got
skipped. There are a few files for the RPC trigger which exceed
this limit. In this commit they are added back in since some
jobs need them to run properly as they are configured to use
the files rather than the database.


Former-commit-id: 0e3b98480984cf14be42b6d51b5935f442cdc498