TrigT1CaloEvent: Const fixes.
A long-standing DataVector bug means that a non-const pointer can be
retrieved from a const_iterator.  Numerous packages ended up relying
on that behavior.  This is another step towards trying to clean this up.

In the map types used by the TrigT1 classes, the pointer value should
be const.  Clients have been adjusted to use the typedefs for the maps;
the change here finally changes the pointers in the maps to be const.



Former-commit-id: a883a6ae83dd4a382d6413cdb329b2bfbfc5b209