Added support for positional numbers wrt given input file in event ranges
This change is for the Event Service. Up intil now the event numbers within
Event Ranges reflected event positions with respect to all input files
provided to the paylod job at startup. This required staging in all inputs
before starting the job.

With the new code, event numbers are treated as positions within the given
input file, which means the inputs can be staged in later, after the job
had already been started.


Former-commit-id: 281bb45ddd16c0080d942d361f47a7959fb83f12