For the ES: cut the output immediately after processing the range
In the new implementation, event range processor fires a dummy
NextEventRange incident immediately after processing given range.
As a result, the range output is finalized and reported to the
scatterer without delay.

Pros:
1. Code simplified
2. This should facilitate the implementation of changes in Pilot
required for interaction with the data prefetcher

Cons:
1. Each worker's run directory contains one extra output file
(create/rewrite is triggered by firing the dummy incident), which
is kind of ugly, but hopefully not a big deal.

Former-commit-id: e0ddfefc810db7ad6e410e788de53f6a61c7e4ab