Fix BMG chamber TDC ticksize
BMG chambers use the HPTDC chip instead of the AMT, so they have a smaller TDC ticksize of .2 ns.  However, the MdtCalibrationSvc was treating all MDT chambers the same.
This change modifies the tdcBinSize() method of MdtCalibrationSvc to check to see if the chamber in question is a BMG chamber and return the correct tick size if so.


Former-commit-id: 0d679a5c10b34e054b1ea8ae62801bb43a5ede4a