Backported the version of AsgTools and AsgExampleTools from the master branch. As some of the analysis tool packages already depend on features that were not part of the code in this branch before.
The one thing that I didn't include from the master branch is the
const-correctness of the standalone implementation of ToolHandle<T>.
Since Gaudi doesn't force const-correctness for its own ToolHandle in
this branch, that just caused a lot of issues for the analysis code.
For no good reason. So the const-correctness issues will only be cleaned
up in the master branch properly.


Former-commit-id: df4428a08cb2c18f129f5eb77d1d2daa7f2f4c40