WIP: AOD reductions for release 21 - ID track pT cut ATLASRECTS-3820
This MR raises the pT threshold for ID TrackParticles from 400 to 500 MeV, by means of the cutLevel in InDetJobProperties (13->14). This is one of the changes agreed by the AOD reduction task force. The change takes effect in the BS->ESD step and is therefore NOT using thinning - the objects never get built.


Former-commit-id: 8347647d2bddab9bfc38d5dee291c3f7156c1729