Added a limit to the number of detail messages in the job report from G4Exceptions (ATLASJT-357)
G4 exceptions can be fatal, therefore, even warnings associated to them are added to the job
report. However, there was no limit on the number of messages that could be added. This has led
to grid tasks producing job report files with over 8MB. The fixes included in this branch place
a limit on the number of messages that can be included in the job report from G4 exceptions.


Former-commit-id: 2d77252c28d65277612a175437b0caed2c297ff4